requirements.txt
Package             Version
------------------- -------
asgiref             3.8.1
Django              5.1
django-cors-headers 4.4.0
djangorestframework 3.15.2
pip                 24.2
python-dotenv       1.0.1
setuptools          60.2.0
sqlparse            0.5.1
tzdata              2024.1
wheel               0.37.1
